REMD rigid body fit

+++ continuousflex/protocols/protocol_genesis.py
@@ -550,43 +550,47 @@ class ProtGenesis(EMProtocol):
                 runParallelJobs(cmds, env=self.getGenesisEnv(), numberOfMpi=numMpiPerFit,
                                 numberOfThreads=self.numberOfThreads.get(), hostConfig=self._stepsExecutor.hostConfig)
 
-                # append files
-                if iterFit != 0:
+                if self.rb_n_iter.get()> 1 :
+                    if self.simulationType.get() == SIMULATION_REMD:
+                        raise RuntimeError("Simulation REMD not allowed for Rigid body fitting iteration > 1")
+
+                    # append files
+                    if iterFit != 0:
+                        for i2 in range(n_parallel):
+                            indexFit = i2 + i1 * numParallelFit
+                            tmpPrefix = self._getExtraPath("%s_tmp" % str(indexFit + 1).zfill(5))
+                            newPrefix = self.getOutputPrefix(indexFit)
+
+                            cat_cmd = "cat %s.log >> %s.log" % (tmpPrefix, newPrefix)
+                            tcl_cmd = "animate read dcd %s.dcd waitfor all\n" % (newPrefix)
+                            tcl_cmd += "animate read dcd %s.dcd waitfor all\n" % (tmpPrefix)
+                            tcl_cmd += "animate write dcd %s.dcd \nexit \n" % newPrefix
+                            with open("%s.tcl" % tmpPrefix, "w") as f:
+                                f.write(tcl_cmd)
+                            cp_cmd = "cp %s.pdb %s.pdb" % (tmpPrefix, newPrefix)
+                            runCommand(cat_cmd)
+                            runCommand(cp_cmd)
+                            runCommand("vmd -dispdev text -e %s.tcl" % tmpPrefix)
+
+                    rstfile = ""
                     for i2 in range(n_parallel):
                         indexFit = i2 + i1 * numParallelFit
-                        tmpPrefix = self._getExtraPath("%s_tmp" % str(indexFit + 1).zfill(5))
                         newPrefix = self.getOutputPrefix(indexFit)
+                        if iterFit != 0:
+                            tmpPrefix = self._getExtraPath("%s_tmp" % str(indexFit + 1).zfill(5))
+                        else:
+                            tmpPrefix = self.getOutputPrefix(indexFit)
 
-                        cat_cmd = "cat %s.log >> %s.log" % (tmpPrefix, newPrefix)
-                        tcl_cmd = "animate read dcd %s.dcd waitfor all\n" % (newPrefix)
-                        tcl_cmd += "animate read dcd %s.dcd waitfor all\n" % (tmpPrefix)
-                        tcl_cmd += "animate write dcd %s.dcd \nexit \n" % newPrefix
-                        with open("%s.tcl" % tmpPrefix, "w") as f:
-                            f.write(tcl_cmd)
-                        cp_cmd = "cp %s.pdb %s.pdb" % (tmpPrefix, newPrefix)
-                        runCommand(cat_cmd)
-                        runCommand(cp_cmd)
-                        runCommand("vmd -dispdev text -e %s.tcl" % tmpPrefix)
+                        runCommand("cp %s.rst %s.tmp.rst" % (tmpPrefix, newPrefix))
+                        rstfile += "%s.tmp.rst "%newPrefix
+                        #save angles
+                        angles = self._getExtraPath("%s_current_angles.xmd" % str(indexFit + 1).zfill(5))
+                        saved_angles = self._getExtraPath("%s_iter%i_angles.xmd" % (str(indexFit + 1).zfill(5), iterFit))
+                        runCommand("cp %s %s" % (angles, saved_angles))
 
-                rstfile = ""
-                for i2 in range(n_parallel):
-                    indexFit = i2 + i1 * numParallelFit
-                    newPrefix = self.getOutputPrefix(indexFit)
-                    if iterFit != 0:
-                        tmpPrefix = self._getExtraPath("%s_tmp" % str(indexFit + 1).zfill(5))
-                    else:
-                        tmpPrefix = self.getOutputPrefix(indexFit)
-
-                    runCommand("cp %s.rst %s.tmp.rst" % (tmpPrefix, newPrefix))
-                    rstfile += "%s.tmp.rst "%newPrefix
-                    #save angles
-                    angles = self._getExtraPath("%s_current_angles.xmd" % str(indexFit + 1).zfill(5))
-                    saved_angles = self._getExtraPath("%s_iter%i_angles.xmd" % (str(indexFit + 1).zfill(5), iterFit))
-                    runCommand("cp %s %s" % (angles, saved_angles))
-
-                    #cleaning
-                    runCommand("rm -rf %s" %self._getExtraPath("%s_tmp" % str(indexFit + 1).zfill(5)))
-                self.inputRST.set(rstfile)
+                        #cleaning
+                        runCommand("rm -rf %s" %self._getExtraPath("%s_tmp" % str(indexFit + 1).zfill(5)))
+                    self.inputRST.set(rstfile)
             self.inputRST.set(initrst)
 
     def createGenesisInputFile(self,inputPDB, outputPrefix, indexFit):
